Add login page with Google Authentication
Implemented a new login page that includes both regular and Google sign-in options. The Vue component contains mechanics for both sending standard username and password data to the server, and for processing Google sign-in credentials, both sessions routing to respective endpoints. This enhancement aids in streamlining login process for users.