Improve authentication process and add logout functionality
Added middleware for login and improved the login process by reloading the Nuxt app after a successful login. A new "Log out" option was also added in the headers, which updates the state of the user to unauthenticated and redirects to the main page. All these changes enhance the user experience by providing a more reliable and intuitive authentication system.