Add car management feature to user profile
Implemented the ability for users to manage their cars. Added new routes to get a list of cars and add new cars for a user. Updated the user and car schema to reflect this relationship. The UI in the profile's page has also been updated to display car data properly.