Create 'apts' Mongoose schema in new file
This commit introduces a new Mongoose schema definition for the 'apts' (appointments) database table in a new file. This schema defines the structure for the appointments, including attributes such as 'appointmentDate', 'carId', 'service' and 'status', amongst others. It will function as the main data model to be used with Nuxt/Mongoose to manage appointments data in the server.