Switch to runtime configuration for sensitive data
This commit replaces hardcoded sensitive data with runtime configurations in the payment, server API, and appointment creation code. Moving this data to a secure and flexible runtime configuration enhances the security of the application and makes it easier to manage and control environment-specific configurations. Changes also include the switch from a sandbox to a production environment.